app/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
decorator==5.0.9
Django==3.2.6
django-debug-toolbar==3.2.1
django-el-pagination==3.3.0
django-extensions==3.1.3
django-silk==4.1.0
djangorestframework==3.12.4
Faker==8.11.0
gprof2dot==2021.2.21
idna==3.2
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
psycopg2==2.9.1
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.9.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
